refactor(sessions): dedupe generated transcript parsing
